% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CruiseCheckeRDBI.R
\name{CruiseCheckeRDBI}
\alias{CruiseCheckeRDBI}
\title{Cruise data check via R}
\usage{
CruiseCheckeRDBI(dat.source = "csv", year = 2018, lake = 2,
  target = 1, stype = 1, username = "", password = "",
  dbname = "")
}
\arguments{
\item{dat.source}{This determines if the function should operate on local csv files or
the results of database queries. If 'database', the user will be prompted
to provide username, password, and dbname, which in the case of the Great
Lakes Science Center, should be GLSC.}

\item{year}{Character vector, year value in quotes. Provides information about the year, which is
used in maming output files.}

\item{lake}{Character vector. For example, "Michigan". Provides information about
location which is used in naming output files.}

\item{target}{Numeric vector representing the study or study type. For example, target = c(209) would
return data collected during the Lake Michigan LWAP fall acoustic survey.}

\item{stype}{Numeric vector representing the type of sample data to be evaluated. For example,
SAMPLE_TYPE (stype) = 1 for trawl data.}

\item{username}{Character vector. This is the username required to access the database, if required.}

\item{password}{Password for database access.}

\item{dbname}{Name of the data source. If using ROracle, it should be GLSC.}
}
\value{
Generates an html file with various plots and checks of data consistency,
  accuracy, and completeness.
}
\description{
Cruise data check via R
}
\details{
The purpose of this function is to create tables and plots to help the user
 evaluate the consistency, completeness, and accuracy of data that are either
 in an Oracle database or are being prepared for appending to an Oracle
 database. Currently the functionality is limited to data and data structures
 that match those of the Research Vessel Catch database developed by the USGS
 Great Lakes Science Center, which is used for storing data collected during
 a wide variety of both field and laboratory sampling exercises. More specifically,
 the current iteration ONLY works with trawl data. Future developments could add
 other sample types.
}
