% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_table_names.R
\name{Get_table_names}
\alias{Get_table_names}
\title{Get Database Table Field Names}
\usage{
Get_table_names(username, password, dbname, sample_type)
}
\arguments{
\item{username}{Character string providing the username required to access the database, if required.}

\item{password}{Character string providing password for database access.}

\item{dbname}{Character string providing the name of the data source. If using ROracle, it should be GLSC.}

\item{sample_type}{Character string naming the sample type}
}
\value{
This is a helper function that gets the names of the fields in tables used
 for trawl data that are in the database. The field names in csv input
 files are compared against these names.
}
\description{
Get Database Table Field Names
}
\details{
This script is a helper function. In its current state, all it does
 is generate Rdata data sets that contain the database field names.
}
