% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRoxDataDoc.R
\name{makeRoxDataDoc}
\alias{makeRoxDataDoc}
\title{Make roxygen document for data included with R packages.}
\usage{
makeRoxDataDoc(dataFrame, title = substitute(dataFrame))
}
\arguments{
\item{dataFrame}{The name of the data frame for which you wish to create roxygen output}

\item{title}{The title for the data frame}
}
\value{
Text in roxygen format for use in documenting data included with R packages.
}
\description{
Make roxygen document for data included with R packages.
}
