% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_names.R
\docType{data}
\name{op_names}
\alias{op_names}
\title{Operations data field names}
\format{data.frame
op_names}
\usage{
data(op_names)
}
\description{
Operations data field names
}
\keyword{datasets}
