% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_catch_names.R
\docType{data}
\name{tr_catch_names}
\alias{tr_catch_names}
\title{Trawl catch data field names}
\format{data.frame
tr_catch_names}
\usage{
data(tr_catch_names)
}
\description{
Trawl catch data field names
}
\keyword{datasets}
