% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_fish_names.R
\docType{data}
\name{tr_fish_names}
\alias{tr_fish_names}
\title{Trawl individual fish data field names}
\format{data.frame
tr_fish_names}
\usage{
data(tr_fish_names)
}
\description{
Trawl individual fish data field names
}
\keyword{datasets}
