% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_l_names.R
\docType{data}
\name{tr_l_names}
\alias{tr_l_names}
\title{Trawl length data field names}
\format{data.frame
tr_l_names}
\usage{
data(tr_l_names)
}
\description{
Trawl length data field names
}
\keyword{datasets}
