% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_lf_names.R
\docType{data}
\name{tr_lf_names}
\alias{tr_lf_names}
\title{Trawl length freqeucny data field names}
\format{data.frame
tr_lf_names}
\usage{
data(tr_lf_names)
}
\description{
Trawl length freqeucny data field names
}
\keyword{datasets}
