% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_op_names.R
\docType{data}
\name{tr_op_names}
\alias{tr_op_names}
\title{Trawl operations data field names}
\format{data.frame
tr_op_names}
\usage{
data(tr_op_names)
}
\description{
Trawl operations data field names
}
\keyword{datasets}
